Fixed the pakname not being set when reading a PAK file causing errors later when trying to retrieve data.
unPAK'ing implemented (though not extensively tested).

Notes:
While unPAK'ing, replacing files is handled badly. For every single duplicate file, it will ask you each time without "Yes to all" or "No to all" options. This may or may not be implemented in the future, but the .NET CLR provides no MessageBox button options for "Yes to all" and "No to all," so a dialog would have to be coded from scratch. Frankly, I'm not sure if it's worth it.